Time series writer perf update
Summary:
The original structure was to have all data in the 'time_series' table.
This didn't scale well and got pretty slow over time. I changed the
table to be partitioned by day, so we could drop older data. This didn't
allow us to filter based on time without doing a table scan, since only
primary keys are honored in the partition.

This new structure fixes us to 30-second intervals in the time series
writing. We pre-write the new times to ts_time, then insert any new
node/key combos which points to a value in ts_value. Next step is some
form of partitioning on the data in ts_value. Maybe that's a HASH
partition on the id, or the time, I'm not sure yet.

Test Plan: Running for the last day. I copied ~10GB of data from the old
time_series tables for a few hours and we have the last 24 hours of data
so this is good enough to flip over. Query times look better, but
certainly not perfect.

Reviewers:

Subscribers:

Tasks:

Blame Revision: