Introduce timestamp to bucket conversion methods in BucketMap
Summary:
This is a first pass for staggering bucket time boundaries. Added bucket, timestamp, buckets, and duration methods. Also added static versions in case we have the shardId/windowSize but not the map itself. Cleaned up most usages of FLAGS_bucket_size in both handlers. The contracts should be that duration is not dependent on the shardId, that each pair of methods is reversible (e.g. bucket->timestamp->bucket), and that the difference in timestamps between two buckets is equal to duration.

Next step is to double check all scheduled threads. Then external calls (rollups, zippyDB, client code). Lastly, switch it on and redo all the tests.

Reviewed By: scottfranklin

Differential Revision: D4548253

fbshipit-source-id: 8ab63f72af8f586d36ae5a6e2487a846432179e3