Check timestamp delta across bucket boundaries
Summary:
Add the minTimestampDelta check for empty bucket scenario except for first timestamp.

Context: TimeSeriesStream uses a parameter named minTimestampDelta to skip successive input timestamps that are too close to each other. appendTimestamp method had a bug that skipped this check if the bucket was empty. The fix addresses this bug but takes care to ensure that we handle the special case for first timestamp.

Reviewed By: jeffdshen

Differential Revision: D4971881

fbshipit-source-id: f9e3776bec3e32050b2aa3dbc23ce26aec9f35d6