Don't corrupt data if we can't read key_list files
Summary:
If we fail to read a key_list file, we end up with block data that is not
mapped to any timeseries.
This would normally be discarded. However, we allow data points to be ingested
before we finish reading block files in order to minimize down time. When this
happens, the data points create new timeseries using "free" ids that happen to
exist in the block files, so when we read the block data it gets attached to the
wrong timeseries.

The fix here is simple. Make a copy of the key list before we allow new points,
then check membership against that instead of the current list.

Reviewed By: jeffdshen

Differential Revision: D5033658

fbshipit-source-id: 6e8b8ddffc816c35cc8869d24587be4872987be6