Improve BeringeiServiceHandler unit tests
Summary:
All the unit tests only checked to see if we put in at least one datapoint. We also only had one datapoint in tests because we tried to put in datapoints that were too far behind, which got adjusted to now (which then blocks putting in datapoints before now).

This fixes both, except in AddDroppedShardBack and ShuffleShards, which return less data than the total amount, most likely because adding back a shard may load non-determinate amounts of data back depending on if the queues were flushed. In those cases, we only check for at least one datapoint, not the exact amount.

Reviewed By: mlgood

Differential Revision: D4919919

fbshipit-source-id: 022cd494472768ce33bcd325f1a47036d99deafd