Add .dockerignore file
Summary:
This is a very simple PR. It just adds a .dockerignore file. This helps with docker caching, since we add the whole directory while building the docker image, ignoring those files does not trigger new steps. As a result cached steps will not be used only when the code itself changes.
Closes https://github.com/facebookincubator/beringei/pull/47

Differential Revision: D5150059

Pulled By: mlgood

fbshipit-source-id: f38952640abbe0d023b3f918cd87a7e0cf6f7b68