Pull scan results from controller and store in DB
With this diff, NMS connects to the controller every 60s and
downloads all scan results, stores them to the database and removes them
from the controller.

Tested with NMS running in a container on my devserver, with a 2 node
setup.
Thrift files generated with `thrift -gen js:node Controller.thrift`