poll latest sjc topology
when the NMS first loads it polls topology from disk
then sends that to BQS for stats type ahead map, which
is used for matching metric short names -> key ids in
beringei. we then update that from the controller. when
this diverges we get an initial period where the stats are
using stale keys from the old topology.

the stats ta map will be moved to poll api service at some
point in the near future to resolve this.