Upgrade babel to 7.0.0-beta.56 (#258)
Summary:
Pull Request resolved: https://github.com/facebookexternal/terragraph-nms/pull/258

Upgrades babel to 7.0.0 -- this is required for things like storybook, and we really should be using
the latest anyway.  I had to make some changes in dependencies/plugins, but mostly things are backward
compatible.

See https://babeljs.io/docs/en/next/v7-migration.html

This matches the version in xplat/js, so it should be well baked.

Reviewed By: apbuteau

Differential Revision: D9384349

fbshipit-source-id: f171d701fa7759a84772055729b86bcc4b314de9