drop reports for new node macs in stats writer, prevent empty node names
Summary:
In /stats_writer we receive a list of node mac -> key/values.
If we haven't seen the mac before we insert a new node record, with an
empty name.
We expect TopologyFetcher to update it later, but this leaves a bunch of
empty node names in the DB that we try to reconcile as duplicate node names.

Two options.
- Ignore empty names from duplicate check.
- Stop recording nodes that don't exist in the topology (which we don't allow you to view anyways).

I'm going with option B. This means we'll be dropping reports until a MAC is
added into the topology.

Reviewed By: csmodlin

Differential Revision: D9501654

fbshipit-source-id: 1d654d9afe7755c53f8159f29ee927fdafa5276f