Support multiple scales per grafana dashboard
Summary: Scales are matched by looking for keyname (i.e. what you search for in the key field) as a substring in the target name (i.e. what is shown as the label for each data series). This is imperfect but works well if you use generic search terms such as "uptime" or "rssi". The imperfectness of this method is explained with an info popover on the scale input. In the case that a scale cannot be matched, it falls back on the first query's scale, which means this won't change the original solution's handling of the scale.

Reviewed By: csmodlin

Differential Revision: D10211486

fbshipit-source-id: 850b8557d2c4dcbb58a4a68cfe487d2ed434ab87