fix site polarity visualization
Summary:
The site view was broken and only showed even/odd.
We piggy-back off node polarity for this view, but there's no such thing as a plain hybrid - so we need to have a special case. I considered using a large int but using a string is fine to separate from the enum space.

Reviewed By: elludraon

Differential Revision: D10861405

fbshipit-source-id: b56fbb0ee8c743643ac5ddb66f51a118247e9976