more un-breaking of beringei
Summary:
When I made the breaking change in D12971674, I made it to an existing build, so it had the existing files to GLOB_RECURSE. Trying this on a new build it fails to GLOB since the OUTPUT isn't specified in add_custom_command.
Changing this to just execute the command instead of expecting specific files generated by running the command fixed it on a new build.

Reviewed By: elludraon

Differential Revision: D13142882

fbshipit-source-id: a32b51f1305ad4a46796b421909467d33d7965e4