changes to show link availability on NMS UI
Summary:
Changes to BQS to support reporting and viewing link availability in NMS UI.  The link is "up" if MCS0 management messages continue every 25.6ms.  But MCS0 can pass at SNR 12dB or more below the threshold for data.  If the link is up, mgmtlinkup counter will increment every 25.6ms

The FW will decide that the link is not available (LINK_UP_DATADOWN state) if the SNR is too low or the PER too high to support data. If the link is available for data, a linkAvailable counter will increment every 25.6ms.  If at any point the linkAvailable counter < mgmtlinkup, it means that for at least a portion of the time interval, the link was not available for data.

This diff calculates the % available and shows times when the link is not available on the availability graph on the UI.

This feature was originally requested by Daniel.

{F144451508}

Reviewed By: pmccut

Differential Revision: D12920723

fbshipit-source-id: 1a711f8c85d6a5d618ef7e7f001e0a863d6f5ba0