Fix writing to DB for TOPO scans
Summary:
When TOPO scans were added, if they are started immediately, the BWGD index is 0.

The scan response service was checking the BWGD index to avoid duplicate writes to the DB.

Rather than writing start_BWGD to the DB, this diff now writes curSfNum / 16.  For all other scan types, curSfNum/16 and start_bwgd are the same thing.

Reviewed By: pmccut

Differential Revision: D13017897

fbshipit-source-id: ebb320f089eb5c58cd65fb10cab1626f634ab60a