link analyzer draft
Summary:
Using sequelize to display the list of tests and the 'health' field. I used
test data that csmodlin provided and added a couple link names and different
health values to show this.

When selecting a test it switches the tile to Monochrome to make it easier to
view, and map tile selection was redone so it doesn't break when setting this
outside the modal.

TODO:
- De-select test results view, I'm thinking we change the top bar so it's easy to de-select
- Only show test results for topology selected (one line)
- Right details box like link analyzer does now
- Table view on the bottom
- Cleanup
- Graphs from mysql once we know the format
- Something showing historical link health from beringei 15min?

Reviewed By: csmodlin

Differential Revision: D10048891

fbshipit-source-id: 12d8658f2f28f6425f3487eebaab6827f052a196