Create python handler to aggregate multi-hop data about the network
Summary:
The motivation behind creating this pipeline is two-fold. One, we would like to schedule iperf and ping sessions to run on segments of the network to assess link health. Two, we would like to view statistics about the network in NMS.

This diff provides the functionality to extrapolate the closest POP(s), the number of wireless hops to get to the POP(s), the number of P2MP hops, whether the POP(s) can be reached in additional equal cost routes (ECMP), and the path to the POP(s) for any given node in the network topology.

In order to determine the correct POPs, we invoke the `getRoutes` endpoint on the controller API. Using `aiohttp` and `asyncio`, we can "parallelize" the API requests by making them asynchronously (i.e. the result of a given (node, POP) is not dependent on the result for any other pair). `getRoutes` returns routes that provide the minimum metric by performing a depth first search in the network. It also guarantees no cycles.

From here, I iterate through the `getRoutes` results and aggregate the desired statistics mentioned above (while removing any POP to POP or ethernet (wired) links).

I created a test topology and mocked the results of network call in order to test some of the desired functionality in the primary function under test, `get_routes_for_nodes`.

Reviewed By: csmodlin

Differential Revision: D13122340

fbshipit-source-id: 8b692a94da241b4e48af7e715dbad1e185775904