Change RouteToPop representation
Summary: Previously, the path attribute of a `RouteToPop` object was a list of wireless hop tuples. Comparing this representation with the output of the `getRoutes` API call to ensure that the route didn't change mid-test would be difficult. After discussing with Cory, Anubhav, and Shardul last week, we decided that saving the `getRoutes` output directly would make more sense.

Reviewed By: anubhavsingla

Differential Revision: D13400967

fbshipit-source-id: e06a3c6e82964aa7a1426af49833a3c7c67ab8f5