bqs: bump writer queue sizes for beringei client
Summary: While testing D12938529 I noticed lots of missing data showing up in random ways. This was due to queue sizes, but the running size flags being used were > # of DPs being set. There's some silly logic in BeringeiClient to assume an average vector size of 500 per write request, which seems like the cause. I set this to 1 to ignore that logic and make the queue size worst case be as large as we're expecting. This has been running for about a day in LAB (~200K active DPs) and no issues thus far. If we have a really large topology and these aren't enough, we could bump the settings OR sleep between write calls.

Reviewed By: drhuanliu

Differential Revision: D13008325

fbshipit-source-id: fd0e8565e21f001df3228e56a0ac77d73685398d