Grafana part I: return keyname with query.
Summary:
This is part I of diffs related to Grafana.  Prior to this diff, if a requested
key has a short name, only the name of the link and direction are returned.
This diff changes that so that the keyname is also returned in the column
names.  The column names are the names displayed on the graph.

Reviewed By: pmccut

Differential Revision: D13588785

fbshipit-source-id: a669492973931754e99be091dc986aa15c5711f9