Bugfix: matching node names per link
Summary:
Grafana has pull downs for nodeA/nodeZ - nodeZ applies only to link stats.  When you select a nodeA, it queries BQS which returns a list of nodes which form links with nodeA.

There were two issues - one is that each link appears only once in the topology with "nodeA" and "nodeZ" representing the two nodes in a random or alphabetical order (doesn't matter here).  So we need to check both nodeA and nodeZ in the topology against the Grafana nodeA.  (we were checking only topology:nodeA)

The second issue was that we weren't limiting to wireless links and Grafana was also showing wired links for which there are no stats.

Differential Revision: D13662748

fbshipit-source-id: 7f610dfe2ad580ce2e23f7117b38e0b96cf2cf82