Script for identifying CN power status
Summary: When computing availability metrics, it is important to discard cases where the subscriber has powered down their CN. The heuristic for determining the power status of a given CN is explained in the comments. We poll 15 minutes worth of stats because E2E attempts beamforming regularly for 30 seconds before entering a "back-off" period that only attempts every 5 minutes for any given link.

Reviewed By: anubhavsingla, csmodlin

Differential Revision: D13640234

fbshipit-source-id: 33eeefc00bf02a4a3a1ccd34771ab0d689ebc3f0