Add restrictors to key search
Summary:
In grafana, you can specify a restrictor and a key.  Prior to this diff, when you select a key, you get a pull-down with possible keys but it did not take the restrictor into account.

In this diff, the typeahead search in BQS now includes the restrictor.

As in other Grafana diffs, files in "dist" are machine-generated; do not review them.  See files in "src"

Reviewed By: pmccut

Differential Revision: D13737023

fbshipit-source-id: f621369f2e453f11ab59dde4a6412b8d0e8e66d1