Add power status script to the insights pipeline
Summary:
We wanted to write the power status of CNs (and eventually all responder nodes) to beringei. I changed get_power_status to write to beringei directly rather than returning a dictionary with the node power statuses. In addition, I changed NodePowerStatus to be a subclass of IntEnum so I could write the value as an enum to the database.

I added a test class that writes the power status stats once and reads them back to make basic sanity checks. So far we have:

1) verify that there is one node power status stat for each CN in each topology
2) verify that each node power status is a valid enum value

Finally, I added a call to generate_insights so get_power_status is now part of the analytics pipeline.

Reviewed By: anubhavsingla

Differential Revision: D13910110

fbshipit-source-id: 9cfa2dbc86b6fc9919068f3021391fddc5e2d014