Apply restrictors during search.
Summary:
This diff fixes an issue with restrictors during a typeahead search and allows a setting to prevent multiple keys with the same keyname from being returned.

Restrictors are either a list of nodes (specified by a name or MAC address) or link names and the returned results are limited to nodes or links that are on the restrictor list.  Before this diff, all keys were searched and then restrictors were applied.  The problem is that the first search stops after `limit` (default 100) results are found and restrictors are applied after the first search which means there could be <<100 results.  In this diff, restrictors are applied in the initial search.

I added some optimizations to the restrictor search but it's still slow (I have ASAN turned on so maybe it's not as bad but it's taking 1-2s per search) because it does a regex search on every single key - for SJC this could be really slow.  I think a search with restrictors is common and in that case, we can greatly speed up the search by organizing the map structures using src MAC address as the key. (for a future diff)

Reviewed By: pmccut

Differential Revision: D13790740

fbshipit-source-id: cba39ac0d5c4c9ae4b838e8542981f1c98aceaf4