Adding support for Speed test
Summary:
Speed test is a different variant of the Multi-hop Network Health Test Plan.
In Speed test, the UI specifies a Pop-Node link that is to be tested.
This is done my passing `pop_to_node_link` parameter in the API call.
The parameter is to be an object of `Pop_to_Node_Link` struct of `NetworkTest.thrift`.

This diff also restructures `views.py` to make the code legible and clean (it was getting a bit clustered).

Reviewed By: omikader

Differential Revision: D14000748

fbshipit-source-id: 0369d79250c698ade069354d307647da41f457d9