Start agg_key id at 1000000000
Summary:
The problem as described in the task is that when reading a stat (e.g. MCS) the data fetched from Beringei was for another unrelated stat.  Anubhav saw this problem on his setup very recently - so we need to fix it now.
The reason for the wrong data is that there are two MySQL tables whose id fields are used as the Beringei ID (ts_key and agg_key) and if these two tables have overlapping ids, then, on write, Beringei will write the aggregate key and not the node key.

At first I tried using a different shard for the two sets of stats but this would require a big change because we use the key id as the "key" in several std::unordered_map objects in the code and assume that the key id is unique.  The problem would be if we queried both an agg_key and a ts_key, the query would fail.
For some reason, the AUTO_INCREMENT command in the cxl.sql file does not work when starting up - no idea why.

Another possibility would have been to combine the ts_key and agg_key tables - also a big change.
So in this solution, the current id is checked before creating/altering the agg_key table and if it is < 100000000, set it to 100000000.

Reviewed By: pmccut

Differential Revision: D14077600

fbshipit-source-id: f52349c00571582ea2926d779311388753f37a00