Add snappy compression and thrift serializer for scan results.
Summary:
This diff modifies scan result writes to MySQL.

BQS polls the E2E for new scan results for all controllers in the topologies table and writes the results to MySQL.

In this diff, the format of ScanResp is changed.  Before this diff, the algorithm was

ScanResp -> simpleJsonSerialize (human readable text) -> MySQL COMPRESS() -> BLOB

this path required us to quantize the double precision results (to avoid SNR = 3.3333333333333 being written to the DB) and to remove fields from ScanResp to avoid overflow.

The new algorithm in this diff is:

ScanResp -> BinarySerialize (thrift serialize) -> Google Snappy compress -> BLOB

Now we don't need to quantize or to remove any fields and we're not storing json keys in the DB.

There is a bug in the thrift implementation of thrift CompactSerializer in node.js - endian for DOUBLE conversion is backwards relative to the c++.  The BinarySerializer does not have this problem so not patch.sh needed if we use BinarySerialize.  Because we are also using compression, it shouldn't matter in terms of size and BinarySerializer is faster.

Reviewed By: noahfren

Differential Revision: D14072498

fbshipit-source-id: 59ecb5a5c070311564e9cf37c786dcd95a903b43