bqs: prometheus /metrics poll endpoint
Summary:
I started testing prometheus by pushing to the prometheus pushgateway, but that was only a trial. Now it's time to write the data the way prometheus expects with a poll.
Push was bad because the metrics don't expire - so we keep the old values in the pushgateway indefinitely. This forces a timestamp onto each metric and once polled it's gone.

Reviewed By: csmodlin

Differential Revision: D14383617

fbshipit-source-id: 5811066de391451642505fca295881beca64bfb4