Add pipeline to detect non-recoverable outage
Summary:
This is a pipeline that runs every 30s to detect and even where the link goes down and stays down after multiple ignition retries.  It works as follows

every 30s:
- read mgmtLinkUp for both ends of the link using 30s stats for 240 seconds (8 samples)
- if, left to right, the link is up for at least 1 sample and right to left, mgmtLinkUp is either 0 or nan for at least 2 samples, the the first condition is met
- read link up attempts 3 times at times 0, 30, 60 - if the E2E has attempted to bring up the link in both 30s intervals at least once, the second condition is met
- if both conditions are met, the event is true and, for now, we increment a counter and report the stat

The next step will be to specify a white list of links for which, if the event is triggered, we will restart minion on both sides of the link.

Reviewed By: anubhavsingla

Differential Revision: D14653393

fbshipit-source-id: 7789c0d0e989b41a65836d0d1d5fbf7fca060b78