Remove checks for non-30s-spaced stats in link up calculation
Summary:
This diff removes checks in the algorithm that computes link up time in BQS.  The checks were looking for non-uniform 30s time buckets.   With Anubhav's recent changes to the stats agent, these checks are no longer necessary.

Fan found a case where the check was erroneously causing the algorithm to detect link down when in fact the link did not go down.  That was because, for unknown reasons, the mgmtLinkUp FW counter decreased over a time interval (see the task) which is not possible in firmware.  (debugging that will be another task)

Reviewed By: pmccut

Differential Revision: D14728011

fbshipit-source-id: 75b1191b057c75547360862251860d8e69650e10