Write rx scan results with empty responses to db
Summary:
In order to calculate the response rate of rx nodes for a network scan, we need to keep track of the nodes that do not respond to the scan, rather than filter them out as we currently do.

This diff writes each rx response as a row in the rx_scan_results table. If the rx scan response had no route info, the scan_resp column for the entry is marked as null.

Reviewed By: csmodlin

Differential Revision: D14432923

fbshipit-source-id: 8e883a1b236e1e523718730f3a9b412b52062b4e