Add ability to restart minion if link ignition fails
Summary:
This diff adds the capability to the insight pipeline to restart minion if an event is hit.

We add a configuration in AnalyticsConfig.json that lets us enable/disable each pipeline and adds a whitelist of links on which to take action (or "all").  The json file is read every 30s so you can make changes without restarting the container.

I also fixed a bug in previous diff the would have caused failure if the topology list changes while the container is up.

Reviewed By: anubhavsingla

Differential Revision: D14681747

fbshipit-source-id: 5248ae142d63a669025e4d479b3239fbd82698bb