Fix buffersize check in the UDP socket creation logic
Summary:
Noticed some strange logging behavior and chatted with Aijay. We found that there was a bad check in the open-source code. This didn't affect the logic or execution of the code (just logging)

The check should be looking to see if the buffer size that was requested is equal to the actual buffer size, not if it is double.

Reviewed By: aijayadams

Differential Revision: D14785361

fbshipit-source-id: 2b95af46024cf08e60be17eb38acbfefb938b590