Fix upgrade view refresh bug
Summary:
NodeUpgradeTable wasn't updating in real-time. The problem was because filteredData was being set once at the beginning (thanks Alex!) Essentially, the new status reports for the nodes were never making it to NodeUpgradeTableEntry). I fixed this issue by moving filteredData out of state and doing the filtering in render. This is slightly less efficient, but perhaps a memoization technique can be later applied if the component becomes too un-performant.

I also re-ran "yarn run thrift-to-flow" to update our client-side typing and added the flow directive to all of the upgrade components.

Reviewed By: aclave1

Differential Revision: D14843153

fbshipit-source-id: 7d2822511924b30580cb522fdc031b7b5506fedf