Flow type Models
Summary:
Add flow types to our models. It's a bit hacky because the types are defined in both sequelize and in flow types, and they need to stay in sync. We could potentially write something to parse it, but it's a big project. If someone adds something without adding types, they'll get flow errors.
If someone deletes something without changing the types, they'll get runtime errors. We can write a test to enforce that the types are correct, but I will look into using ES6 classes with sequelize.
Also had to modify the sequelize flow types - findOne and create return an instance with the types on them. I believe that the correct way to use the library is to use the `.get(...)` function as opposed to directly accessing the variables, but I don't want to change a ton of code. Open to alternatives.

**I'm looking for comments on how this is implemented.** Maybe we keep the code fixes and don't go this route? Maybe something else?

Having done this, I have uncovered a number of issues:
- verificationType is undefined in the model but used otherwise.
- role vs superUser uses in scripts was incorrect

Reviewed By: aclave1

Differential Revision: D14721187

fbshipit-source-id: beddcbad5fb98cec7e863d6463f9f2ac2694908a