fix overlay strategy crash
Summary: There are 2 overlay strategies, test and metric. Each strategy has a list of overlays such as health, mcs, etc. If an overlay is selected (like RSSI), and the user changes betwen strategies (selects a test or exits a test), and the overlay is not part of the new strategy, there will be an inconsistency where the selected overlay is not part of the new strategy and a null reference exception is thrown. This fixes that in a few places.

Reviewed By: omikader

Differential Revision: D14858754

fbshipit-source-id: 8d0d2a45c7f38013e3358c150227d00c9802b9b3