fix unauthorized ajax request redirect
Summary:
Whenever an unauthenticated request is received, the useragent is instructed to redirect to the login page. The redirect contains a url parameter with the original url, that way the user is redirected back to their originally intended url after signin. This is good for direct browsing / link clicking, but bad for ajax. The old behavior redirected the useragent to the ajax url. After login, the user would be redirected to /topology/list for example (and the user would see JSON after logging in).

The correct behavior is for the axios interceptor to override the return url with the useragent's current location (window.location.href).

Reviewed By: pmccut

Differential Revision: D14880563

fbshipit-source-id: aa8b1e9213e1318fa3e4707c377e15f31fd322de