Use getDefaultRoutes API endpoint in network test
Summary:
D13744032 added the ability to get the default routes (i.e. shortest routes to the outside network) directly from the controller. Previously, getting the default routes for multi-hop test was a naive approach in which the routes for every node x pop_node pair were computed, and then iterated over to determine which one had the fewest wireless hops.

Now, the `defaultRoutes` endpoint provides us with a guarantee that the routes returned will all be the shortest paths (in terms of number of wireless hops) to the external network. This saves us time, as we only need to make 1 API call to the controller instead of node x pop_node `getRoute` calls.

This change also allows us to remove `aiohttp` from the dependency requirements and overall simplifies the logic for `get_routes_for_nodes` tremendously.

**getDefaultRoutes was added in M35. Any invocations of get_routes_for_nodes in versions prior to that will simply return an empty list**

Reviewed By: spurav

Differential Revision: D14952492

fbshipit-source-id: 4557951637e36e8d3693f11c92dcc14575352389