Flow type user portion of request
Summary:
Realized that we never updated the request portion after adding flow types in sequelize.
Fixing that now.

This resulted in fixing the user create/update - the inputs were never validated.
Not a big deal since the endpoint is authenticated, but it's a good idea.

Also remove `verificationType` from the field map - it's never used anywhere:
{F156441595}

Reviewed By: rckclmbr

Differential Revision: D15057205

fbshipit-source-id: 9d349e6a6c48903a46bfd654fcfee2b69cc2813c