Speed up ZMQ requests
Summary:
As each iperf/ping job was submitted over ZMQ, the db was also being updated.  The db update took around 200ms and this meant that on a network with hundreds of links, the job submission took a couple of minutes and the test timed out before all iperf/ping sessions were complete.  On SJC, for example, only 100 out of >500 links reported output - the others timed out.

This diff moves the db update out of the job submission loop - now it takes on the order of 1ms per job submission (250x speed-up).

Also in this diff, updated logging to include timestamp.  And moved the print of the iperf/ping results to log level debug.

Reviewed By: spurav

Differential Revision: D15063675

fbshipit-source-id: 21f083d995cd17afdb1e77c07484794f3b9cf89b