Keep writing results until ZMQ Queue is empty
Summary:
mysql writes are taking around 50ms per update.  In SJC, there are around 225 links (550 link directions) and write both iperf and ping results -> 550 * 50ms * 2 (iperf + ping) ==> around 55s.

The timeout is set to 15s; if the writes go past 15s, the thread controlling writes is killed and we lose the results.

In this diff, the thread controlling writes dies after the 15s timeout AND when the ZMQ queue is empty.

Also, simplified the db update() but that didn't make a large difference in optimizing the time.

Reviewed By: spurav

Differential Revision: D15072434

fbshipit-source-id: d115e2b6559b201c07bd2ec9c9ce10d9b26e72db