Write UDP pinger stats directly to Prometheus and update default config values
Summary:
This change adds functionality to support communicating with Prometheus directly and removes the logic that wrote pinger stats to Beringei. The new logic spins up an HTTP endpoint and uses the Prometheus singleton to write stats that are flushed out at the end of each request.

In addition, I have changed the default `num_packets` and `ping_interval_s` from 200 --> 20 and 3s --> 1s, respectively.

Reviewed By: csmodlin

Differential Revision: D15174456

fbshipit-source-id: 2d90fe9106205c36528a499ff22ad05e3e6ad7d4