Report event when optimizer detects anomaly
Summary:
When the auto optimizer detects link re-ignition, it restarts minion
and, in this diff, reports an event.  If the optimizer detects re-ignition but
not does restart minion because of back-off timer, it still reports an event
but with INFO level instead of WARNING

Event is written directly to BQS.   With Kafka running, we can change this.

Reviewed By: anubhavsingla

Differential Revision: D14970593

fbshipit-source-id: 72bed76fcf63326abfef9a8816e17f0a48a08961