Calculate network scan response rate
Summary:
For a given window of time on a network, calculate the valid response rate of scans of the given type (e.g. IM, PBF)

The stats are broken down by scan group, which is a group of scan responses that share the same initation (either a scheduled scan or a manualy triggered scan).

Stats for a given scan group fall into two categories: tx_stats which contains info about tx node responses and rx_stats which contains info about rx node responses.

We consider well-formed tx responses with a status of zero valid. If we do not recieve a tx response from a node but we recieve rx responses that point to the unresponsive node as a tx node, we create a placeholder tx response, which we consider invalid.

We consider well-formed rx responses with a status code of zero and a non-empty routeInfoList valid.

Note: The CLI is a temporary endpoint for this analysis, we will be writing the stats into the scans db eventually.

Reviewed By: csmodlin

Differential Revision: D14809378

fbshipit-source-id: bb0307b77858f95d302007e1c4775b73b42e5921