Calculate per link scan state
Summary:
For each link, it may be helpful to know at what point in the scan cycle the connection is at. (i.e. Is the current beam a result of PBF, RTCAL, etc.) This diff queries scan data and the topology to find the most recent scan type for each direction involved in a link.

Note: The CLI endpoint is only tmporary, we will write these stats back to the scan db in the future and display them in more useful places like NMS or Grafana.

Reviewed By: fansunfb

Differential Revision: D14809550

fbshipit-source-id: 8d20e022dcfde62c213eeda22f93b84fabadf97f