Improve performance of UDP pinger
Summary:
This primarily changes the UDP pinger main process from using `folly::eventBase` + `folly::AsyncTimeout` to `std::thread` for running periodic timers.

folly's event loop construct is just a single threaded infinite cycle that makes one task at a time and adds it to its queue to run. That being said, I don't see any reason why we would want the pinger to block on behalf of fetching updated test plans, or computing aggregated stats. This diff spins up three threads for performing the following tasks:

1. Fetching the latest topology config and creating a list of test plans
2. Using the list of test plans to ping
3. Aggregating the latest 30s worth of results to create rolled up stats

This diff changes the logging verbosity in UdpPinger to reduce the noise in the logs. I also change the other consumers of API service to use the returned object directly after confirming it is not folly::none instead of saving the value in a separate variable.

Reviewed By: csmodlin

Differential Revision: D15349921

fbshipit-source-id: 7ea388cb5321ecb773ddbd9e7f4b64826d7f306f