Scheduler
Summary:
Add a scheduler to the network test.  Replace Django http server with flask.  The reason for that is to allow the entire network test to run in the same process.

The scheduler lets the user schedule network tests using a cron-like scheduler.  If, for example, you would like to run the parallel test every day at 4:30am and 6:30am, you would enter

cron_minute = 30
cron_hour = 4,6
other cron entries (month, day of week, day of month) = "*"

or, if you'd like to run it now, there is an "ASAP" field - set the field to "1" and then it will ignore other fields.

Lastly, there is also a priority field - the scheduler uses python's "sched" which honors the priority.  ASAP is always highest priority.

Different from the current way it works, if a test is running and you schedule a new test ASAP, it will queue it, not discard it.  But this only allows one ASAP test at a time.

Also different from the current test, everything is per network (as opposed to global).  There is one scheduler thread running per network.

The schedule survives a container restart by reading the db at the very beginning looking for any scheduled tests.

The diagram below shows the updated block diagram of network test.

{F157692896}

Reviewed By: spurav

Differential Revision: D15226580

fbshipit-source-id: 73a98b7d07fc72aa8f6eb0a900cfa6080ee8d05f