Update Event.thrift in NMS
Summary: `LOGTAIL` wasn't showing up as a possible Event table filter. Despite having a copy of `Event.thrift` with it present, the problem was that it was pulling from `gen_nodejs` instead of `shared/types`. I fixed this and also copied over the latest version of `Event.thrift` from `META-TERRAGRAPH`.

Reviewed By: elludraon

Differential Revision: D15495923

fbshipit-source-id: 13bfc2824a6acf90a19e068fb5e845301a1bcae8