Patch for Network Test Scheduler
Summary: This is a patch so that we can upgrade network test using the scheduler until Alex is back and can address some issues in D15384459.   With this patch, we can run tests, see results, and abort tests.  We need to wait for D15384459 to be able to see scheduled tests and modify the schedule (and to make it less hacky).

Reviewed By: pmccut

Differential Revision: D15461731

fbshipit-source-id: c0a14ddceb63d6ec4e38e6b5e58d1245101a22f8