Add link health map overlay
Summary:
Current "health" overlay is actually the link ignition state - not the health.  This diff adds a health overlay that reads the health calculated by analytics.

Requires 1s (high frequency) stats to work and requires the short name "link_health" to be added.

Update:
Numbers are removed; unknown is gray instead of maroon.

Reviewed By: elludraon

Differential Revision: D15425125

fbshipit-source-id: e28f04a62b14ce5852296321290ba66246c8cb48