Add scan group_id to tx_scan_response table and scan_response_rate table
Summary: As of D15075980, controller generates an id assigned to scan groups (a group of scan responses that are results of the same scan scheduling event) which we will use for analysis of scan data. Add this field to tx_scan_response so we can use it more easily for scan response analysis.

Reviewed By: fansunfb

Differential Revision: D15720638

fbshipit-source-id: 3acaa31e0d2173af9d3474fdbbda15bd7f75c60f