Add scan group id to scan results db
Summary:
As of D15075980, controller generates an id assigned to scan groups (a group of scan responses that are results of the same scan scheduling event) which we will use for analysis of scan data. Add it to BQS and network_scan.

Note: the group id is reset on controller restart so we will need to use it in combination with timestamps to determine scan groups

Reviewed By: fansunfb

Differential Revision: D15720643

fbshipit-source-id: 0fc823e2cd4cc4b52a0997ec2634115d08f84530