Read from Prometheus instead of Beringei
Summary:
Change analytics stats reading from Beringei to Prometheus.

This code continues to use numpy_time_series as before with small changes to the API.

numpy_time_series now calls prometheus_time_series to read from Prometheus instead of from Beringei.  Stats writes are still to Beringei (changing soon).

For each time series, we need to read both the data and the timestamp (two separate reads).  This is the only way to recover the input timestamp and remove unknown samples from the timeseries.

Only "short names" are supported -- stats for which the keyname is constant across all links/nodes in the network.

This code uses asyncio but only at the numpy_time_series level.  It needs to be optimized to bring asyncio another level up so truly parallelize the reads.

Update:
Output of prometheus time series is now
- times - unique timestamps
- values - corresponding values

Output is in numpy format (I think it's more convenient).  Functions to computes bucket indices are moved to the prometheus time series class because they are db-specific.

Reviewed By: anubhavsingla

Differential Revision: D15599074

fbshipit-source-id: 1aaf16aa63969dfd2631858b53aea1020045f5c4