Split inventory to server and UI components (#357)
Summary:
Pull Request resolved: https://github.com/facebookincubator/magma/pull/357

Pull Request resolved: https://github.com/facebookexternal/fbc/pull/1002

Pull Request resolved: https://github.com/facebookexternal/terragraph-nms/pull/264

Having server and UI exist in the same project has caused quite a few problems.
This is an attempt to separate them to different packages.  This will allow us
to code and configure them in isolation, remove UI source from production builds,
and (possibly) replace the server component with something of another language in
the future.

Would love comments on how to improve it beyond this implementation.

Reviewed By: aclave1

Differential Revision: D15868790

fbshipit-source-id: 1b3ec312fac0422c8b0a01764fae4d58bb7885b4