save translations to docker volume
Summary:
Saves locale files to a persistent docker volume. The alternative to this method would be saving to the DB. I'm not against saving to the db but this was easier and faster.

Nokia requires an api to post translation files to. Originally, these were saved to the local nms filesystem. This was an oversight as the files can be deleted if the container is deleted/restarted.

Reviewed By: rckclmbr

Differential Revision: D15610510

fbshipit-source-id: 5d71610fc11d1fcd22ad240a26297ec3384a89ff