Drop unused arguments from PrometheusUtils writeMetrics function
Summary:
I noticed that the `topologyName` and `jobName` parameters were unused in `writeMetrics` when reviewing D15501566. This diff removes those fields and simplifies the `UdpPingClient` logic by no longer grouping pinger results by network name and instead writing them all to the same metrics list.

To reduce code duplication for `writeAggrResults` when creating the metrics, I made an assumption (which I believe to be fairly safe) that a nodeName and network/topology name on the same NMS would never be the same.

Reviewed By: pmccut

Differential Revision: D15849801

fbshipit-source-id: 0c50c9368b7346a537fea1b2bcc7b110fbadcd9d