Change start and end time in scan_response_rate table to use bwgd
Summary:
MySql timestamps did not provide us enough precision to accurately track which scans have already been analyzed as part of a scan_response_rate table entry so this diff updates the table to use bwgd instead.
This table is still unused in production so this change to the table should not require migrations.

Reviewed By: csmodlin

Differential Revision: D16014580

fbshipit-source-id: 773a2df3c85f9923d4d7938692fc8f2887175230