Add relay lint rules (#1091)
Summary:
Pull Request resolved: https://github.com/facebookexternal/fbc/pull/1091

After D16086703 (and the corresponding fix D16095716), we noticed that we have these nicely generated relay flow types, and instead we're using manually hand-typed ones (that are wrong) and don't get updated. This leads to a ton of bugs like specifying fields that don't exist, specifying unused fields, coding mistakes, etc.

This enables lint for that.
This is a HUGE problem.

Reviewed By: apbuteau

Differential Revision: D16097812

fbshipit-source-id: 22557131419b497cab412546cd3ae77899057c8d